backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
httpx==0.27.0
redis==5.0.8
python-dotenv==1.0.1
mistralai==1.0.0
elevenlabs==1.5.0
pydantic-settings==2.0.0
